+++ src/main/java/com/iexec/core/replicate/ReplicatesService.java
@@ -235,27 +235,19 @@ public class ReplicatesService {
         ChainContribution chainContribution = null;
         String resultLink = null;
         String chainCallbackData = null;
-        TaskDescription taskDescription = null;
+        final TaskDescription taskDescription = iexecHubService.getTaskDescription(chainTaskId);
 
-        switch (statusUpdate.getStatus()) {
-            case CONTRIBUTED:
-                workerWeight = iexecHubService.getWorkerWeight(walletAddress);
-                chainContribution = iexecHubService.getChainContribution(chainTaskId, walletAddress).orElse(null);
-                break;
-            case RESULT_UPLOADED:
-                ReplicateStatusDetails details = statusUpdate.getDetails();
-                if (details != null) {
-                    resultLink = details.getResultLink();
-                    chainCallbackData = details.getChainCallbackData();
-                }
-                taskDescription = iexecHubService.getTaskDescription(chainTaskId);
-                break;
-            case COMPUTED:
-            case RESULT_UPLOAD_FAILED:
-                taskDescription = iexecHubService.getTaskDescription(chainTaskId);
-                break;
-            default:
-                break;
+        if (statusUpdate.getStatus() == CONTRIBUTED || statusUpdate.getStatus() == CONTRIBUTE_AND_FINALIZE_DONE) {
+            workerWeight = iexecHubService.getWorkerWeight(walletAddress);
+            chainContribution = iexecHubService.getChainContribution(chainTaskId, walletAddress).orElse(null);
+        }
+
+        if (statusUpdate.getStatus() == RESULT_UPLOADED || statusUpdate.getStatus() == CONTRIBUTE_AND_FINALIZE_DONE) {
+            final ReplicateStatusDetails details = statusUpdate.getDetails();
+            if (details != null) {
+                resultLink = details.getResultLink();
+                chainCallbackData = details.getChainCallbackData();
+            }
         }
 
         return UpdateReplicateStatusArgs.builder()
@@ -427,12 +419,12 @@ public class ReplicatesService {
             return Either.left(error);
         }
 
-        if (newStatus == CONTRIBUTED) {
+        if (newStatus == CONTRIBUTED || newStatus == CONTRIBUTE_AND_FINALIZE_DONE) {
             replicate.setContributionHash(updateReplicateStatusArgs.getChainContribution().getResultHash());
             replicate.setWorkerWeight(updateReplicateStatusArgs.getWorkerWeight());
         }
 
-        if (newStatus == RESULT_UPLOADED) {
+        if (newStatus == RESULT_UPLOADED || newStatus == CONTRIBUTE_AND_FINALIZE_DONE) {
             replicate.setResultLink(updateReplicateStatusArgs.getResultLink());
             replicate.setChainCallbackData(updateReplicateStatusArgs.getChainCallbackData());
         }